L2-836: CR changes
