+++ frontend/src/tests/lib/api/sns.api.spec.ts
@@ -3,12 +3,7 @@
  */
 
 import type { HttpAgent } from "@dfinity/agent";
-import {
-  AccountIdentifier,
-  ICP,
-  LedgerCanister,
-  type SnsWasmCanisterOptions,
-} from "@dfinity/nns";
+import { ICP, LedgerCanister, type SnsWasmCanisterOptions } from "@dfinity/nns";
 import { Principal } from "@dfinity/principal";
 import mock from "jest-mock-extended/lib/Mock";
 import { get } from "svelte/store";
@@ -18,7 +13,6 @@ import {
   querySnsSwapCommitment,
   querySnsSwapState,
   querySnsSwapStates,
-  querySwapCanisterAccount,
 } from "../../../lib/api/sns.api";
 import {
   importInitSnsWrapper,
@@ -167,13 +161,5 @@ describe("sns-api", () => {
 
     expect(ledgerCanisterMock.transfer).toBeCalled();
     expect(notifyParticipationSpy).toBeCalled();
-  });
-
-  it("should return swap canister account", async () => {
-    const expectedAccount = await querySwapCanisterAccount({
-      swapCanisterId: rootCanisterIdMock,
-      controller: mockIdentity.getPrincipal(),
-    });
-    expect(expectedAccount).toBeInstanceOf(AccountIdentifier);
   });
 });
