+++ frontend/src/tests/lib/services/sns.services.spec.ts
@@ -3,12 +3,7 @@ import { Principal } from "@dfinity/principal";
 import * as api from "../../../lib/api/sns.api";
 import * as services from "../../../lib/services/sns.services";
 import { mockMainAccount } from "../../mocks/accounts.store.mock";
-import {
-  mockIdentity,
-  mockPrincipal,
-  resetIdentity,
-  setNoIdentity,
-} from "../../mocks/auth.store.mock";
+import { mockIdentity, mockPrincipal } from "../../mocks/auth.store.mock";
 import { mockSnsSwapCommitment } from "../../mocks/sns-projects.mock";
 
 const { participateInSwap, getSwapAccount } = services;
@@ -81,23 +76,10 @@ describe("sns-services", () => {
   });
 
   describe("getSwapAccount", () => {
-    const spyQuery = jest
-      .spyOn(api, "querySwapCanisterAccount")
-      .mockImplementation(() =>
-        AccountIdentifier.fromHex(mockMainAccount.identifier)
-      );
     afterEach(() => jest.clearAllMocks());
     it("should return the swap canister account identifier", async () => {
       const account = await getSwapAccount(mockPrincipal);
       expect(account).toBeInstanceOf(AccountIdentifier);
-      expect(spyQuery).toBeCalled();
-    });
-    it("should return undefined if no identity", async () => {
-      setNoIdentity();
-      const account = await getSwapAccount(mockPrincipal);
-      expect(account).toBeUndefined();
-      expect(spyQuery).not.toBeCalled();
-      resetIdentity();
     });
   });
 });
